requirements.txt
discord-py-interactions==5.10.0
python-dotenv==1.0.1